usm: tls: tests: Move helpers
